+++ packages/components/accordion/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Patch Changes
 
+- Updated dependencies []:
+  - @jala-banyu/divider@3.1.0
 - Add figma design to storybook
 
 ## 1.0.1